Rename "session media elements" to "participating media elements"
This is (hopefully) a more memorable name, and is in analogy to the
"slaved media elements" of media controllers.

Just like slaved media elements, the set does not need to be maintained
explicitly, since it's defined by which media elements have a media
session as its current media session, so it's always in sync.

This leaves the concept unused, but it won't be for long.