Separate HTMLMediaElement-based invocation/interruption/release of media sessions from the definition of media sessions.
This is the first step to making media session objects more generic. We still need to decide how to replace 'participating media elements' and 'active participating media elements' within the generic media session algorithms with something that takes a better primitive (e.g. a base AudioContext).

This is a work in progress. Feedback welcome.