Ensure that only the audio-producing objects that were playing, or otherwise running, when a media session was interrupted are resumed at the end of that interruption.
Instead of saying that all 'audio-producing participants' should duck/unduck or pause/unpause (depending on the type of interruption) we should be clear that if there are e.g. 4 audio-producing participants attached to a media session but only 1 of those audio-producing objects is actually playing/running that only that object should be resumed.