Remove MediaSession.setMetadata() pending likely breaking changes
There are a few open issues around metadata:
https://github.com/whatwg/mediasession/issues/58
https://github.com/whatwg/mediasession/issues/70
https://github.com/whatwg/mediasession/issues/76

In particular the need for feature detection means that we must have
an interface, which will be a breaking change. Remove the code until
this is more clear, to not tempt anyone into implementing it.

Much of this, in particular the prose and example, can be revived.