Remove the MediaRemoteControls interface pending flattening
There are a few open issues around remote controls:
https://github.com/whatwg/mediasession/issues/11
https://github.com/whatwg/mediasession/issues/21
https://github.com/whatwg/mediasession/issues/42
https://github.com/whatwg/mediasession/issues/45
https://github.com/whatwg/mediasession/issues/56

In particular issue 45 will change this API a lot, so in order to not
tempt implementors into going with this design, remove it for now.

Much of this, in particular the prose and example, can be revived
once we have the time to hammer out the flattened API design.